FEAT: Improved http_handling
